requirements.txt
mysqlclient
django
requests
bs4
lxml
django_apscheduler
ujson
psutil
djangorestframework
markdown
django-filter
